Add tests for the directives controller
Currently, our tests don't cover the directives controller. This commit
expands existing service tests to cover the controller, and adds a few
additional test cases.
