+++ packages/api/src/validators/shared.test.ts
@@ -1,5 +1,6 @@
 import {
   validateBodyFromAuthentication,
+  validateBodyFromAdminAuthentication,
   fieldsFromUserOrAdminAuthentication,
 } from "./shared";
 
@@ -86,6 +87,98 @@ describe("validateBodyFromAuthentication", () => {
       validateBodyFromAuthentication({
         emailFromAuthToken: "test@permanent.org",
         userSubjectFromAuthToken: "not_a_uuid",
+      });
+    } catch (err) {
+      error = err;
+    } finally {
+      expect(error).not.toBeNull();
+    }
+  });
+});
+
+describe("validateBodyFromAdminAuthentication", () => {
+  test("should find no errors in valid parameter set", () => {
+    let error = null;
+    try {
+      validateBodyFromAdminAuthentication({
+        emailFromAuthToken: "test@permanent.org",
+        adminSubjectFromAuthToken: "b2a6787c-f255-465a-8eb0-1583004d4a4f",
+      });
+    } catch (err) {
+      error = err;
+    } finally {
+      expect(error).toBeNull();
+    }
+  });
+  test("should raise an error if emailFromAuthToken is missing", () => {
+    let error = null;
+    try {
+      validateBodyFromAdminAuthentication({
+        adminSubjectFromAuthToken: "b2a6787c-f255-465a-8eb0-1583004d4a4f",
+      });
+    } catch (err) {
+      error = err;
+    } finally {
+      expect(error).not.toBeNull();
+    }
+  });
+  test("should raise an error if emailFromAuthToken is the wrong type", () => {
+    let error = null;
+    try {
+      validateBodyFromAdminAuthentication({
+        emailFromAuthToken: 1,
+        adminSubjectFromAuthToken: "b2a6787c-f255-465a-8eb0-1583004d4a4f",
+      });
+    } catch (err) {
+      error = err;
+    } finally {
+      expect(error).not.toBeNull();
+    }
+  });
+  test("should raise an error if emailFromAuthToken is an invalid value", () => {
+    let error = null;
+    try {
+      validateBodyFromAdminAuthentication({
+        emailFromAuthToken: "not_an_email",
+        adminSubjectFromAuthToken: "b2a6787c-f255-465a-8eb0-1583004d4a4f",
+      });
+    } catch (err) {
+      error = err;
+    } finally {
+      expect(error).not.toBeNull();
+    }
+  });
+  test("should raise an error if adminSubjectFromAuthToken is missing", () => {
+    let error = null;
+    try {
+      validateBodyFromAdminAuthentication({
+        emailFromAuthToken: "test@permanent.org",
+      });
+    } catch (err) {
+      error = err;
+    } finally {
+      expect(error).not.toBeNull();
+    }
+  });
+  test("should raise an error if adminSubjectFromAuthToken is the wrong type", () => {
+    let error = null;
+    try {
+      validateBodyFromAdminAuthentication({
+        emailFromAuthToken: "test@permanent.org",
+        adminSubjectFromAuthToken: 1,
+      });
+    } catch (err) {
+      error = err;
+    } finally {
+      expect(error).not.toBeNull();
+    }
+  });
+  test("should raise an error if adminSubjectFromAuthToken is the wrong format", () => {
+    let error = null;
+    try {
+      validateBodyFromAdminAuthentication({
+        emailFromAuthToken: "test@permanent.org",
+        adminSubjectFromAuthToken: "not_a_uuid",
       });
     } catch (err) {
       error = err;
